Fixed two bugs: 1. A typo in the total calculation for total_pitNoSlab in pathogenflows.R. 2. A problem in the emissions.R script where HUMAN_DATA$popurb and HUMAN_DATA$poprur were used for the per capita emissions calculations. Instead HUMAN_DATA$population should have been used together with HUMAN_DATA$fraction_urban_pop.
One minor problem remains. For some of the island states (Comoros, Djibouti, Guinea Bissou, Kiribati, Maldives, Sri Lanka, Solomon Islands, Saint Vincent and the Grenadines, Vanuatu and Samoa) the urban emissions have not been plotted on the map. The reason for this is that the gridded urban and rural population files indicate only rural population on these islands. That is due to a preprocessing script. As these islands do not contribute much to the overall picture, I decided to ignore this problem right now.